Title: 가장 많이 등장한 부분 문자열, Time: 305ms, Memory: 18MB, Status: Passed - CodeTree

+++ 241108/가장 많이 등장한 부분 문자열/most-common-substring.java
@@ -12,7 +12,7 @@ public class Main {
 		String S = st.nextToken();
 
 		int p = 31, m = (int) 1e9 + 7;
-		int[] pPow = new int[L + 1];
+		long[] pPow = new long[L + 1];
 		pPow[0] = 1;
 		for (int i = 1; i <= L; i++) {
 			pPow[i] = (pPow[i - 1] * p) % m;
